CONTRIB-8660: Move recording_helper to appropriate classes
None of this behaviour belongs outside of the appropriate classes.

* All fetches for the recording object belong to the persistent.
* The proxy should be responsible for sorting and parsing of its own data
