+++ classes/broker.php
@@ -79,4 +79,51 @@ class broker {
         // Everything is valid.
         return null;
     }
+
+    /**
+     * Helper for responding when recording ready is performed.
+     *
+     * @param instance $instance
+     * @param array $params
+     */
+    public static function recording_ready(instance $instance, array $params): void {
+        // Decodes the received JWT string.
+        try {
+            $decodedparameters = JWT::decode(
+                $params['signed_parameters'],
+                config::get('shared_secret'),
+                array('HS256')
+            );
+        } catch (Exception $e) {
+            $error = 'Caught exception: ' . $e->getMessage();
+            header('HTTP/1.0 400 Bad Request. ' . $error);
+            return;
+        }
+
+        // Validations.
+        if (!isset($decodedparameters->record_id)) {
+            header('HTTP/1.0 400 Bad request. Missing record_id parameter');
+            return;
+        }
+
+        $recording = recording::get_record(['recordingid' => $decodedparameters->record_id]);
+        if (!isset($recording)) {
+            header('HTTP/1.0 400 Bad request. Invalid record_id');
+            return;
+        }
+
+        // Sends the messages.
+        try {
+            // We make sure messages are sent only once.
+            if ($recording->get('status') != recording::RECORDING_STATUS_NOTIFIED) {
+                notifier::notify_recording_ready($instance->get_instance_data());
+                $recording->set('status', recording::RECORDING_STATUS_NOTIFIED);
+                $recording->update();
+            }
+            header('HTTP/1.0 202 Accepted');
+        } catch (Exception $e) {
+            $error = 'Caught exception: ' . $e->getMessage();
+            header('HTTP/1.0 503 Service Unavailable. ' . $error);
+        }
+    }
 }
